[TIMOB-26340] Android: Update V8 runtime to 6.9.x - V8 6.9.427.17 and NDK r16b (#9926)
* [TIMOB-26340] Android: Update V8 runtime to 7.0.x
* Use V8 7.0.276.9
* Build with NDK r16b.
* Fix compilation issue around use of args.Callee(), replace with different hack to grab URL. Make note that we may not need this hack at all.
* Bump to 8.0.0
* bump android module api version
* Use playservices/map/touchid modules on android built against an SDK using v8 7.0
* Support file URIs for pre-packaged modules - particularly so we can test with newly re-compiled native modules locally when revving V8
* Upload rebuilt native modules and update sha hashes for them
* Use custom build of ios ti.facebook module without breaking bug
* Move away from deprecated APIs and be more explicit in error handling.
* Axe the undocumented 'inherit' method to extend Titanium proxy types.
* Clean up cpp proxy template some: Move from placing the java class name into a special __javaClass__ property off the contructor of proxies to wrapping the jclass in an External and setting it as the Data for the FunctionTemplate. Makes things easier/simpler to grab the java class needed to instantiate instances through JNI when constructing Java-backed proxies.
* Move timeout of 20 minutes just around the test command, use timeout of 5 minutes for clone of test suite.
* Comment out LDLIBS instructions that break native modules

+++ build/utils.js
@@ -138,6 +138,15 @@ function cachedDownloadPath(url) {
 }
 
 Utils.generateSSRIHashFromURL = function (url, callback) {
+	if (url.startsWith('file://')) {
+		// Generate integrity hash!
+		ssri.fromStream(fs.createReadStream(url.slice(7))).then(integrity => {
+			callback(null, integrity.toString());
+		}).catch(e => {
+			callback(e);
+		});
+		return;
+	}
 	const downloadPath = cachedDownloadPath(url);
 	fs.removeSync(downloadPath);
 	download(url, downloadPath, function (err, file) {
@@ -154,11 +163,23 @@ Utils.generateSSRIHashFromURL = function (url, callback) {
 };
 
 Utils.downloadURL = function downloadURL(url, integrity, callback) {
-	const downloadPath = cachedDownloadPath(url);
+	if (url.startsWith('file://')) {
+		if (!fs.existsSync(url.slice(7))) {
+			return callback(new Error('File URL does not exist on disk: %s', url));
+		}
+		// if it passes integrity check, we're all good, return path to file
+		ssri.checkStream(fs.createReadStream(url.slice(7)), integrity).then(() => {
+			// cached copy is still valid, integrity hash matches
+			callback(null, url.slice(7));
+		}).catch(e => callback(e));
+		return;
+	}
 
 	if (!integrity) {
 		return callback(new Error('No "integrity" value given for %s, may need to run "node scons.js modules-integrity" to generate new module listing with updated integrity hashes.', url));
 	}
+
+	const downloadPath = cachedDownloadPath(url);
 	// Check if file already exists and passes integrity check!
 	if (fs.existsSync(downloadPath)) {
 		// if it passes integrity check, we're all good, return path to file
